otii_automation_part/requirements_device.txt
bcrypt==4.1.2
cffi==1.15.1
crc==3.0.1
cryptography==42.0.5
future==1.0.0
ifcfg==0.24
iso8601==2.1.0
paramiko==3.4.0
pycparser==2.21
PyNaCl==1.5.0
pyserial==3.5
PyYAML==6.0.1
scp==0.14.5
tomli==2.0.1
requests==2.32.3
pyroute2==0.8.1
RPi.GPIO==0.7.1